+++ .goreleaser.yaml
@@ -6,7 +6,7 @@ before:
     # You may remove this if you don't use go modules.
     - go mod tidy
 builds:
-  - dir: .
+  - dir: ./cmd
     env:
       - CGO_ENABLED=0
     goos:
