+++ cmd/main.go
@@ -1,19 +1,22 @@
 // Copyright (C) 2024 The Dagu Authors
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package cmd
+package main
 
 import (
+	"os"
+
 	"github.com/dagu-org/dagu/internal/build"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 var (
-	// cfgFile parameter
+	// version is set at build time
+	version = "0.0.0"
+
 	cfgFile string
 
-	// rootCmd represents the base command when called without any subcommands
 	rootCmd = &cobra.Command{
 		Use:   build.Slug,
 		Short: "YAML-based DAG scheduling tool.",
@@ -21,11 +24,28 @@ var (
 	}
 )
 
-// Execute adds all child commands to the root command and sets flags
-// appropriately. This is called by main.main(). It only needs to happen
-// once to the rootCmd.
-func Execute() error {
-	return rootCmd.Execute()
+func main() {
+	if err := rootCmd.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
+
+func init() {
+	build.Version = version
+
+	registerCommands()
+
+	rootCmd.PersistentFlags().
+		StringVar(
+			&cfgFile, "config", "",
+			"config file (default is $HOME/.config/dagu/admin.yaml)",
+		)
+
+	cobra.OnInitialize(func() {
+		if cfgFile != "" {
+			viper.SetConfigFile(cfgFile)
+		}
+	})
 }
 
 func registerCommands() {
@@ -39,23 +59,4 @@ func registerCommands() {
 	rootCmd.AddCommand(schedulerCmd())
 	rootCmd.AddCommand(retryCmd())
 	rootCmd.AddCommand(startAllCmd())
-}
-
-func init() {
-	rootCmd.PersistentFlags().
-		StringVar(
-			&cfgFile, "config", "",
-			"config file (default is $HOME/.config/dagu/admin.yaml)",
-		)
-
-	cobra.OnInitialize(initialize)
-
-	registerCommands()
-}
-
-func initialize() {
-	if cfgFile != "" {
-		viper.SetConfigFile(cfgFile)
-		return
-	}
 }
