+++ cmd/scheduler.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2024 The Dagu Authors
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package cmd
+package main
 
 import (
 	"log"
